requirements.txt
# Core AI/ML Dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
spacy>=3.6.0
opencv-python>=4.8.0
pytesseract>=0.3.10
paddleocr>=2.6.0
ultralytics>=8.0.0

# PII Detection and NLP
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy-transformers>=1.3.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# CLI and Utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
typer>=0.9.0

# Security and Authentication
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
schedule>=1.2.0
pyjwt>=2.8.0

# Workflow Orchestration
apache-airflow>=2.6.0
celery>=5.3.0

# File Processing
pypdf2>=3.0.0
pymupdf>=1.23.0
reportlab>=4.0.0
pillow>=10.0.0
python-magic>=0.4.27

# Image Processing (additional)
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Logging and Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
